Move ResultsTableHeader to components
