+++ src/app/components/site/cs/RoutesCS.tsx
@@ -9,7 +9,7 @@ import {SetQuizzes} from "../../pages/quizzes/SetQuizzes";
 import {MyQuizzes} from "../../pages/quizzes/MyQuizzes";
 import {QuizDoAssignment} from "../../pages/quizzes/QuizDoAssignment";
 import {QuizAttemptFeedback} from "../../pages/quizzes/QuizAttemptFeedback";
-import {QuizTeacherFeedback} from "../../pages/assignment_progress/QuizTeacherFeedback";
+import {QuizTeacherFeedback} from "../../pages/quizzes/QuizTeacherFeedback";
 import {QuizPreview} from "../../pages/quizzes/QuizPreview";
 import {QuizDoFreeAttempt} from "../../pages/quizzes/QuizDoFreeAttempt";
 import {TeacherAccountSelfUpgrade} from "../../pages/TeacherAccountSelfUpgrade";
